Correctly implemented graphicQuality()
Implemented changing the graphic quality correctly. In order to prevent
memory leaks, the entire contents of the spectrogram analysis must be
dumped. In theory, results on the canvas should be unaffected. Pass in
the number of pixels connected to an analyser. (the higher the number,
the lower quality the image).
Added a significant amount of comments to give clarity to what is
happening.
NOTE: I have been unable to test whether there is image loss on the
canvas when switching as I do not have a user input to the graphic
level, I've only been able to trigger a graphic change as a process from
the initialization code in spectrogram.js
*Thus ends the ridiculously long commit log.